Cue-Archived-Project/requirements.txt
Django==1.7.5
django-enumfield==1.2.1
django-recaptcha==1.0.3
django-resized==0.3.5
django-taggit==0.17.0
django-tinymce==2.0.4
enum34==1.0.4
Pillow==2.9.0
psycopg2==2.6
pytz==2015.4
